chore: refactor granjas layout structure
